Commonise the fixtures in the ingests indexer tests
